+++ homeassistant/components/tag/strings.json
@@ -1,7 +1,7 @@
 {
   "title": "Tag",
   "entity": {
-    "event": {
+    "tag": {
       "tag": {
         "state_attributes": {
           "tag_id": {
